+++ recordforwarder/src/send_request_to_lambda.py
@@ -7,23 +7,22 @@ from clients import lambda_client
 from utils_for_record_forwarder import invoke_lambda
 from constants import Constants
 
+CREATE_LAMBDA_NAME = os.getenv("CREATE_LAMBDA_NAME")
+UPDATE_LAMBDA_NAME = os.getenv("UPDATE_LAMBDA_NAME")
+DELETE_LAMBDA_NAME = os.getenv("DELETE_LAMBDA_NAME")
 
-def send_create_request(fhir_json: dict, supplier: str, file_key: str, row_id: str) -> str:
-    """Sends the create request and handles the response. Returns the imms_id."""
+
+def send_create_request(fhir_json: dict, supplier: str, file_key: str, row_id: str):
+    """Sends the create request."""
     # Send create request
     headers = {"SupplierSystem": Constants.IMMS_BATCH_APP_NAME, "BatchSupplierSystem": supplier, "Filename": file_key,
                "MessageId": row_id}
     payload = {"headers": headers, "body": fhir_json}
-    status_code, body, headers = invoke_lambda(lambda_client, os.getenv("CREATE_LAMBDA_NAME"), payload)
-    if status_code != 200:
-        raise MessageNotSuccessfulError(get_operation_outcome_diagnostics(body))
-
-    # Return imms id (default to None if unable to find the id)
-    return "200" or None
+    invoke_lambda(lambda_client, CREATE_LAMBDA_NAME, payload)
 
 
-def send_update_request(fhir_json: dict, supplier: str, file_key: str, row_id: str) -> str:
-    """Obtains the imms_id, sends the update request and handles the response. Returns the imms_id."""
+def send_update_request(fhir_json: dict, supplier: str, file_key: str, row_id: str):
+    """Obtains the imms_id, sends the update request."""
     # Obtain imms_id and version
     try:
         imms_id, version = get_imms_id_and_version(fhir_json)
@@ -39,15 +38,11 @@ def send_update_request(fhir_json: dict, supplier: str, file_key: str, row_id: s
     headers = {"SupplierSystem": Constants.IMMS_BATCH_APP_NAME, "BatchSupplierSystem": supplier, "E-Tag": version,
                "Filename": file_key, "MessageId": row_id}
     payload = {"headers": headers, "body": fhir_json, "pathParameters": {"id": imms_id}}
-    status_code, body, _ = invoke_lambda(lambda_client, os.getenv("UPDATE_LAMBDA_NAME"), payload)
-    if status_code != 200:
-        raise MessageNotSuccessfulError(get_operation_outcome_diagnostics(body))
-
-    return imms_id
+    invoke_lambda(lambda_client, UPDATE_LAMBDA_NAME, payload)
 
 
-def send_delete_request(fhir_json: dict, supplier: str, file_key: str, row_id: str) -> str:
-    """Sends the delete request and handles the response. Returns the imms_id."""
+def send_delete_request(fhir_json: dict, supplier: str, file_key: str, row_id: str):
+    """Obtains the imms_id, sends the delete request."""
     # Obtain imms_id
     try:
         imms_id, _ = get_imms_id_and_version(fhir_json)
@@ -60,25 +55,10 @@ def send_delete_request(fhir_json: dict, supplier: str, file_key: str, row_id: s
     headers = {"SupplierSystem": Constants.IMMS_BATCH_APP_NAME, "BatchSupplierSystem": supplier, "Filename": file_key,
                "MessageId": row_id}
     payload = {"headers": headers, "body": fhir_json, "pathParameters": {"id": imms_id}}
-    status_code, body, _ = invoke_lambda(lambda_client, os.getenv("DELETE_LAMBDA_NAME"), payload)
-    if status_code != 204:
-        raise MessageNotSuccessfulError(get_operation_outcome_diagnostics(body))
-
-    return imms_id
+    invoke_lambda(lambda_client, DELETE_LAMBDA_NAME, payload)
 
 
-def get_operation_outcome_diagnostics(body: dict) -> str:
-    """
-    Returns the diagnostics from the API response. If the diagnostics can't be found in the API response,
-    returns a default diagnostics string
-    """
-    try:
-        return body.get("issue")[0].get("diagnostics")
-    except (AttributeError, IndexError):
-        return "Unable to obtain diagnostics from API response"
-
-
-def send_request_to_lambda(message_body: dict) -> str:
+def send_request_to_lambda(message_body: dict):
     """
     Sends request to the Imms API (unless there was a failure at the recordprocessor level). Returns the imms id.
     If message is not successfully received and accepted by the Imms API raises a MessageNotSuccessful Error.
@@ -94,4 +74,4 @@ def send_request_to_lambda(message_body: dict) -> str:
 
     # Send request to Imms FHIR API and return the imms_id
     function_map = {"CREATE": send_create_request, "UPDATE": send_update_request, "DELETE": send_delete_request}
-    return function_map[operation_requested](fhir_json=fhir_json, supplier=supplier, file_key=file_key, row_id=row_id)
+    function_map[operation_requested](fhir_json=fhir_json, supplier=supplier, file_key=file_key, row_id=row_id)
