+++ recordforwarder/src/utils_for_record_forwarder.py
@@ -2,6 +2,7 @@
 
 import os
 import json
+from errors import MessageNotSuccessfulError
 
 
 def get_environment() -> str:
@@ -22,6 +23,17 @@ def extract_file_key_elements(file_key: str) -> dict:
     return file_key_elements
 
 
+def get_operation_outcome_diagnostics(body: dict) -> str:
+    """
+    Returns the diagnostics from the API response. If the diagnostics can't be found in the API response,
+    returns a default diagnostics string
+    """
+    try:
+        return body.get("issue")[0].get("diagnostics")
+    except (AttributeError, IndexError):
+        return "Unable to obtain diagnostics from API response"
+
+
 def invoke_lambda(lambda_client, lambda_name: str, payload: dict) -> tuple[int, dict, str]:
     """
     Uses the lambda_client to invoke the specified lambda with the given payload.
@@ -39,4 +51,8 @@ def invoke_lambda(lambda_client, lambda_name: str, payload: dict) -> tuple[int, 
         response = lambda_client.invoke(
             FunctionName=lambda_name, InvocationType="Event", Payload=json.dumps(payload)
         )
-        return 200, None, None
+        body = json.loads(response.get("body", "{}"))
+        if response["statusCode"] != "200":
+            raise MessageNotSuccessfulError(get_operation_outcome_diagnostics(body))
+        else:
+            return "200", None, None
